+++ CHANGELOG.md
@@ -9,9 +9,14 @@
 - Implemented blacklist for canonical url
 - Implemented canonical url support for mounte pages, pointing to real page instead of mount path (disabled by default)
 - Implemented expiry date for sitemap entries (customizable with SetupTS or Connector)
+- Implemented pagetitle caching (if there is any USER_INT on the current page)
 - Fixed many bugs and issues
 - Fixed coding style (added .editorconfig)
 - Refactored whole extension
+
+## Beta features
+
+- If you have any issues with cached pagetitle: set `plugin.metaseo.pageTitle.caching = 0` to disable this feature.
 
 ### Migrate from 1.x to 2.x